app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
setuptools==69.5.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
